move fields to own folder, add validation to tags and statuses
