+++ api/models/fields/__init__.py
@@ -1,0 +1,1 @@
+from .color_field import ColorField
